+++ src/main/java/es/geeko/repository/TematicaRepository.java
@@ -3,5 +3,5 @@ package es.geeko.repository;
 import es.geeko.model.Tematica;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TematicaRepository extends JpaRepository<Tematica,Integer> {
+public interface TematicaRepository extends JpaRepository<Tematica,Long> {
 }
